Smart shards Merge

+++ makefile
@@ -80,7 +80,7 @@ OBJS = $(PROJECT_DIR)/main.o $(PROJECT_DIR)/$(ALGFILE)/$(ALGFILE).o $(PROJECT_DI
 
 # extra debug and release flags
 release: CXXFLAGS += -O3 -s -std=c++17
-debug: CXXFLAGS += -O0 -g  -D_GLIBCXX_DEBUG -std=c++17
+debug: CXXFLAGS += -O0 -g -D_GLIBCXX_DEBUG -std=c++17
 
 clang: CXX := clang++
 clang: CXXFLAGS += -std=c++17
@@ -101,11 +101,19 @@ $(EXE): $(OBJS)
 $(PROJECT_DIR)/%.o: $(PROJECT_DIR)/%.c
 	$(CXX) $(CXXFLAGS) -c $< -o $@
 
+# Define a helper function to check dmesg for errors
+define check_failure
+    echo "Make target '$@' failed! Checking dmesg..."; \
+    dmesg | tail -20 | grep -i -E 'oom|killed|segfault|error' || echo "No relevant logs found."
+endef
+
 run: all
-	./$(EXE) $(INPUTFILE)
+	./$(EXE) $(INPUTFILE); exit_code=$$?; \
+	if [ $$exit_code -ne 0 ]; then $(call check_failure); exit $$exit_code; fi
 
 run_debug: debug
-	./$(EXE) $(INPUTFILE)
+	gdb --ex "set print thread-events off" --ex run --ex backtrace --args ./$(EXE) $(INPUTFILE); exit_code=$$?; \
+	if [ $$exit_code -ne 0 ]; then $(call check_failure); exit $$exit_code; fi
 
 # in the future this could be generalized to go through every file in "Tests"
 rand_test: $(PROJECT_DIR)/Tests/randtest.cpp $(PROJECT_DIR)/Common/Distribution.cpp
@@ -128,7 +136,8 @@ test_%:
 	@$(CXX) $(CXXFLAGS) -c -o quantas/$(ALGFILE)/$(ALGFILE).o quantas/$(ALGFILE)/$(ALGFILE).cpp
 	@$(CXX) $(CXXFLAGS) -c -o quantas/Common/Distribution.o quantas/Common/Distribution.cpp
 	@$(CXX) $(CXXFLAGS)  quantas/main.o quantas/$(ALGFILE)/$(ALGFILE).o quantas/Common/Distribution.o -o $(EXE)
-	@./$(EXE) quantas/$(ALGFILE)/$*Input.json
+	@./$(EXE) quantas/$(ALGFILE)/$*Input.json; exit_code=$$?; \
+	if [ $$exit_code -ne 0 ]; then $(call check_failure); exit $$exit_code; fi
 	@$(RM) quantas/$(ALGFILE)/*.o
 	@echo $(ALGFILE) successful
